+++ include/libfork/algorithm/scan.hpp
@@ -9,18 +9,19 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <concepts>   // for invocable
-#include <functional> // for identity, invoke
-#include <iterator>   // for iter_difference_t, random_access_iterator
-#include <ranges>     // for begin, end, iterator_t, range_difference_t
+#include <concepts>    // for same_as
+#include <functional>  // for identity, invoke
+#include <iterator>    // for random_access_iterator, sized_sentinel_for
+#include <ranges>      // for begin, end, iterator_t, random_access_range
+#include <type_traits> // for conditional_t
 
 #include "libfork/algorithm/constraints.hpp" // for indirectly_scannable, projected
-#include "libfork/core/control_flow.hpp"     // for call, fork, join
-#include "libfork/core/invocable.hpp"
-#include "libfork/core/just.hpp"  // for just
-#include "libfork/core/macro.hpp" // for LF_STATIC_CALL, LF_STATIC_CONST, LF_CATCH_ALL
-#include "libfork/core/tag.hpp"
-#include "libfork/core/task.hpp" // for task
+#include "libfork/core/control_flow.hpp"     // for call, dispatch, fork, join
+#include "libfork/core/invocable.hpp"        // for async_invocable
+#include "libfork/core/just.hpp"             // for just
+#include "libfork/core/macro.hpp"            // for LF_STATIC_CALL, LF_STATIC_CONST, LF_ASSERT
+#include "libfork/core/tag.hpp"              // for tag, eager_throw_outside, sync_outside
+#include "libfork/core/task.hpp"             // for task
 
 /**
  * @file scan.hpp