+++ include/libfork/core/exception.hpp
@@ -9,7 +9,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <exception>
+#include <exception> // for exception
 
 #include "libfork/core/first_arg.hpp" // for quasi_pointer