+++ include/libfork/core/impl/unique_frame.hpp
@@ -1,10 +1,12 @@
 #ifndef A7699F23_E799_46AB_B1E0_7EA36053AD41
 #define A7699F23_E799_46AB_B1E0_7EA36053AD41
 
-#include <memory>
+#include <coroutine> // for coroutine_handle
+#include <memory>    // for unique_ptr
 
-#include "libfork/core/impl/frame.hpp"
-#include "libfork/core/macro.hpp"
+#include "libfork/core/impl/frame.hpp"   // for frame
+#include "libfork/core/impl/utility.hpp" // for non_null
+#include "libfork/core/macro.hpp"        // for LF_STATIC_CALL
 
 /**
  * @file unique_frame.hpp