+++ include/libfork/core/just.hpp
@@ -9,20 +9,22 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <concepts>    // for invocable
+#include <concepts>    // for constructible_from, invocable
 #include <coroutine>   // for suspend_never
 #include <exception>   // for rethrow_exception
 #include <functional>  // for invoke
-#include <type_traits> // for invoke_result_t
+#include <type_traits> // for decay_t, invoke_result_t, true_type, false_type
 #include <utility>     // for forward
 
 #include "libfork/core/eventually.hpp"      // for try_eventually
+#include "libfork/core/first_arg.hpp"       // for async_function_object
 #include "libfork/core/impl/awaitables.hpp" // for call_awaitable
 #include "libfork/core/impl/combinate.hpp"  // for combinate
 #include "libfork/core/impl/frame.hpp"      // for frame
+#include "libfork/core/impl/utility.hpp"    // for immovable, unqualified
 #include "libfork/core/invocable.hpp"       // for async_invocable, async_result_t
 #include "libfork/core/macro.hpp"           // for LF_STATIC_CALL, LF_STATIC_CONST, LF_DEPRECAT...
-#include "libfork/core/tag.hpp"             // for tag
+#include "libfork/core/tag.hpp"             // for tag, none
 #include "libfork/core/task.hpp"            // for returnable
 
 /**
@@ -58,11 +60,12 @@ class [[nodiscard("co_await this!")]] just_awaitable : just_awaitable_base<R>, c
  /**
   * @brief Construct a new just awaitable binding the return address to an internal member.
   */
-  template <typename Fun, typename... Args>
-  explicit just_awaitable(Fun &&fun, Args &&...args)
+  template <async_function_object F, typename... Args>
+    requires async_invocable<F, Args...>
+  explicit just_awaitable(F &&fun, Args &&...args)
       : call_awaitable{
             {}, 
-            combinate<tag::call, modifier::none>(&this->ret, std::forward<Fun>(fun))(std::forward<Args>(args)...)
+            combinate<tag::call, modifier::none>(&this->ret, std::forward<F>(fun))(std::forward<Args>(args)...)
         } 
       {}
 
@@ -107,7 +110,6 @@ struct [[nodiscard("co_await this!")]] just_wrapped : std::suspend_never {
       return std::move(val);
     }
   }
-
   /**
    * @brief The value to be forwarded.
    */
@@ -120,13 +122,30 @@ struct [[nodiscard("co_await this!")]] just_wrapped : std::suspend_never {
 template <>
 struct just_wrapped<void> : std::suspend_never {};
 
+namespace detail {
+
+template <class>
+struct some_just_impl : std::false_type {};
+
+template <class T>
+struct some_just_impl<just_awaitable<T>> : std::true_type {};
+
+template <class T>
+struct some_just_impl<just_wrapped<T>> : std::true_type {};
+
+} // namespace detail
+
+/**
+ * @brief Test if a type is a ``just_awaitable`` or ``just_wrapped`` specialization.
+ */
+template <class T>
+concept some_just = detail::some_just_impl<std::remove_cvref_t<T>>::value;
+
 /**
  * @brief A wrapper that supplies an async function with a call operator.
  */
-template <typename F>
-  requires (!std::is_reference_v<F>)
+template <unqualified F>
 struct [[nodiscard("This should be immediately invoked!")]] call_just {
-
   /**
    * @brief Make an awaitable that will call the async function then immediately join.
    */
@@ -135,7 +154,6 @@ struct [[nodiscard("This should be immediately invoked!")]] call_just {
   auto operator()(Args &&...args) && -> just_awaitable<async_result_t<F, Args...>> {
     return just_awaitable<async_result_t<F, Args...>>(std::move(fun), std::forward<Args>(args)...);
   }
-
   /**
    * @brief Immediately invoke a regular function and wrap the result in an awaitable class.
    */
@@ -149,7 +167,6 @@ struct [[nodiscard("This should be immediately invoked!")]] call_just {
       return {{}, std::invoke(std::move(fun), std::forward<Args>(args)...)};
     }
   }
-
   /**
    * @brief The async or regular function.
    */
@@ -163,22 +180,28 @@ struct bind_just {
   /**
    * @brief Make an async function object immediate callable.
    *
+   * We use `std::decay_t` here as `F` may be a reference to function pointer.
+   *
    * @return A functor, that will return an awaitable (in an ``lf::task``), that will trigger a call + join.
    */
   template <typename F>
-  LF_DEPRECATE_CALL LF_STATIC_CALL auto operator()(F fun) LF_STATIC_CONST->call_just<F> {
-    return {std::move(fun)};
+    requires std::constructible_from<std::decay_t<F>, F>
+  LF_DEPRECATE_CALL LF_STATIC_CALL auto operator()(F &&fun) LF_STATIC_CONST->call_just<std::decay_t<F>> {
+    return {std::forward<F>(fun)};
   }
 
 #if defined(__cpp_multidimensional_subscript) && __cpp_multidimensional_subscript >= 202211L
   /**
    * @brief Set a void return address for an asynchronous function.
    *
+   * We use `std::decay_t` here as `F` may be a reference to function pointer.
+   *
    * @return A functor, that will return an awaitable (in an ``lf::task``), that will trigger a call + join.
    */
   template <typename F>
-  LF_STATIC_CALL auto operator[](F fun) LF_STATIC_CONST->call_just<F> {
-    return {std::move(fun)};
+    requires std::constructible_from<std::decay_t<F>, F>
+  LF_STATIC_CALL auto operator[](F &&fun) LF_STATIC_CONST->call_just<std::decay_t<F>> {
+    return {std::forward<F>(fun)};
   }
 #endif
 };