+++ include/libfork/core/tag.hpp
@@ -9,7 +9,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include <type_traits>
+#include <type_traits> // for true_type, false_type
 
 #include "libfork/core/impl/utility.hpp"